requirements.txt
# Core Framework
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0

# API Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-multipart>=0.0.9

# Database
asyncpg>=0.29.0
pgvector>=0.3.0
sqlalchemy>=2.0.0

# ML & NLP
sentence-transformers>=3.0.0
scikit-learn>=1.5.0
numpy>=1.26.0
textstat>=0.7.3

# Structured Output
toon-py>=0.1.0

# OCR (Note: Chandra may need special installation)
# chandra-ocr>=1.0.0  # Uncomment when available

# Cache & Storage
redis>=5.0.0
boto3>=1.34.0

# Utilities
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=24.0.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
httpx>=0.27.0

# Development
black>=24.0.0
ruff>=0.6.0
mypy>=1.11.0

